+++ git-commit.json
@@ -1,1 +1,1 @@
-{"gitCommit":"eb50297"}
+{"gitCommit":"56c2f3a"}
