+++ src/components/Header.tsx
@@ -23,7 +23,7 @@ const Header: React.FC = () => {
       data-module="sdn-header"
     >
       <div className="sdn-header__container govuk-width-container ">
-        {isLandingPage ? logo : <Link href="/domov">{logo}</Link>}
+        {isLandingPage ? logo : <Link href="/">{logo}</Link>}
       </div>
     </header>
   )
