+++ src/pages/domov.module.css
@@ -1,5 +1,5 @@
 .postponeBox {
   background: #f4f4f4;
-  margin-top: 20px;
+  margin-top: 30px;
   padding: 10px 20px 30px 20px;
 }
